+++ test/slo_logoutrequest_test.rb
@@ -281,11 +281,13 @@ class RubySamlTest < Minitest::Test
         logout_request.settings.idp_entity_id = 'https://app.onelogin.com/saml/metadata/SOMEACCOUNT'
         assert logout_request.send(:validate_issuer)
       end
+
       it "return false when the issuer of the Logout Request does not match the IdP entityId" do
         logout_request.settings.idp_entity_id = 'http://idp.example.com/invalid'
         assert !logout_request.send(:validate_issuer)
         assert_includes logout_request.errors, "Doesn't match the issuer, expected: <#{logout_request.settings.idp_entity_id}>, but was: <https://app.onelogin.com/saml/metadata/SOMEACCOUNT>"
       end
+
       it "raise when the issuer of the Logout Request does not match the IdP entityId" do
         logout_request.settings.idp_entity_id = 'http://idp.example.com/invalid'
         logout_request.soft = false
@@ -400,7 +402,7 @@ class RubySamlTest < Minitest::Test
         assert_equal(CGI.unescape(query), CGI.unescape(original_query))
         # Make normalised signature based on our modified params.
         sign_algorithm = XMLSecurity::BaseDocument.new.algorithm(settings.security[:signature_method])
-        signature = settings.get_sp_key.sign(sign_algorithm.new, query)
+        signature = settings.get_sp_signing_key.sign(sign_algorithm.new, query)
         params['Signature'] = Base64.encode64(signature).gsub(/\n/, "")
         # Construct SloLogoutrequest and ask it to validate the signature.
         # It will do it incorrectly, because it will compute it based on re-encoded
@@ -435,7 +437,7 @@ class RubySamlTest < Minitest::Test
         assert_equal(CGI.unescape(query), CGI.unescape(original_query))
         # Make normalised signature based on our modified params.
         sign_algorithm = XMLSecurity::BaseDocument.new.algorithm(settings.security[:signature_method])
-        signature = settings.get_sp_key.sign(sign_algorithm.new, query)
+        signature = settings.get_sp_signing_key.sign(sign_algorithm.new, query)
         params['Signature'] = Base64.encode64(signature).gsub(/\n/, "")
         # Construct SloLogoutrequest and ask it to validate the signature.
         # Provide the altered parameter in its raw URI-encoded form,
@@ -474,7 +476,7 @@ class RubySamlTest < Minitest::Test
         sign_algorithm = XMLSecurity::BaseDocument.new.algorithm(
           settings.security[:signature_method]
         )
-        signature = settings.get_sp_key.sign(sign_algorithm.new, query)
+        signature = settings.get_sp_signing_key.sign(sign_algorithm.new, query)
         params['Signature'] = downcased_escape(Base64.encode64(signature).gsub(/\n/, ""))
 
         # Then parameters are usually unescaped, like we manage them in rails
@@ -552,7 +554,6 @@ class RubySamlTest < Minitest::Test
         logout_request_sign_test.settings = settings
         assert !logout_request_sign_test.send(:validate_signature)
         assert_includes logout_request_sign_test.errors, "Invalid Signature on Logout Request"
-
       end
     end
   end